Agents-Meetings/agents/room_manager_agent/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-openai>=0.1.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-http>=1.21.0
python-dotenv>=1.0.0

